+++ RedPandaIDE/settings.h
@@ -653,8 +653,8 @@ public:
         bool showCodeIns() const;
         void setShowCodeIns(bool newShowCodeIns);
 
-        bool clearWhenEditorHidden();
-        void setClearWhenEditorHidden(bool newClearWhenEditorHidden);
+        //bool clearWhenEditorHidden();
+        //void setClearWhenEditorHidden(bool newClearWhenEditorHidden);
 
         int minCharRequired() const;
         void setMinCharRequired(int newMinCharRequired);
@@ -684,7 +684,7 @@ public:
         int mMinCharRequired;
         bool mHideSymbolsStartsWithTwoUnderLine;
         bool mHideSymbolsStartsWithUnderLine;
-        bool mClearWhenEditorHidden;
+        //bool mClearWhenEditorHidden;
         bool mShareParser;
 
         // _Base interface