+++ RedPandaIDE/settingsdialog/editorcodecompletionwidget.cpp
@@ -26,6 +26,7 @@ EditorCodeCompletionWidget::EditorCodeCompletionWidget(const QString& name, cons
     ui(new Ui::EditorCodeCompletionWidget)
 {
     ui->setupUi(this);
+    ui->chkClearWhenEditorHidden->setVisible(false);
 }
 
 EditorCodeCompletionWidget::~EditorCodeCompletionWidget()
@@ -50,7 +51,7 @@ void EditorCodeCompletionWidget::doLoad()
     ui->chkIgnoreCases->setChecked(pSettings->codeCompletion().ignoreCase());
     ui->chkAppendFunc->setChecked(pSettings->codeCompletion().appendFunc());
     ui->chkShowCodeIns->setChecked(pSettings->codeCompletion().showCodeIns());
-    ui->chkClearWhenEditorHidden->setChecked(pSettings->codeCompletion().clearWhenEditorHidden());
+//    ui->chkClearWhenEditorHidden->setChecked(pSettings->codeCompletion().clearWhenEditorHidden());
     ui->chkHideSymbolsStartWithTwoUnderline->setChecked(pSettings->codeCompletion().hideSymbolsStartsWithTwoUnderLine());
     ui->chkHideSymbolsStartWithUnderline->setChecked(pSettings->codeCompletion().hideSymbolsStartsWithUnderLine());
 
@@ -78,7 +79,7 @@ void EditorCodeCompletionWidget::doSave()
     pSettings->codeCompletion().setShowCodeIns(ui->chkShowCodeIns->isChecked());
     pSettings->codeCompletion().setMinCharRequired(ui->spinMinCharRequired->value());
 
-    pSettings->codeCompletion().setClearWhenEditorHidden(ui->chkClearWhenEditorHidden->isChecked());
+    //pSettings->codeCompletion().setClearWhenEditorHidden(ui->chkClearWhenEditorHidden->isChecked());
 
     pSettings->codeCompletion().setHideSymbolsStartsWithTwoUnderLine(ui->chkHideSymbolsStartWithTwoUnderline->isChecked());
     pSettings->codeCompletion().setHideSymbolsStartsWithUnderLine(ui->chkHideSymbolsStartWithUnderline->isChecked());