+++ NEWS.md
@@ -22,6 +22,8 @@ Red Panda C++ Version 2.24
   - fix: Wrong code suggestion while inputing numbers in assembly files.
   - fix: Defines in all files are wrongly cleared when reparsing.
   - change: New file created by file template is set as unmodified by default.
+  - change: Remove option "clear all symbols when current editor is hidden".
+  - fix: When opening multiple files, only the active file should be parsed.
 
 Red Panda C++ Version 2.23